# swagger.yaml
openapi: 3.0.0
info:
  title: Volunteer Platform API
  version: 1.0.0
  description: Public API docs for donor, volunteer registration & food request flows
servers:
  - url: http://localhost:3000
paths:
  /feedhope-donation-food:
    get:
      summary: Render food donation page (View only)
      responses:
        '200':
          description: Donation page rendered (HTML)
    post:
      summary: Submit food donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - donor
                - foodtype
                - quantity
                - location
                - description
              properties:
                donor:
                  type: string
                foodtype:
                  type: string
                quantity:
                  type: object
                  required: [amount, unit, description]
                  properties:
                    amount:
                      type: number
                    unit:
                      type: string
                    description:
                      type: string
                location:
                  type: string
                description:
                  type: string
      responses:
        '302':
          description: Redirects to nearby requests
...
  /connect-donor-with-requester:
    post:
      summary: Connect donor with requester (Internal use)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - donoor
              properties:
                donoor:
                  type: string
      responses:
        '302':
          description: Redirect to request detail after linking
...